+++ scripts/run_ple2_addressable_eval.py
@@ -84,7 +84,7 @@ def split_sequences(seqs, train_frac, seed):
     rng = random.Random(seed)
     indices = list(range(len(seqs)))
     rng.shuffle(indices)
-    n_train = max(1, int(round(len(seqs) * train_frac)))
+    n_train = max(1, round(len(seqs) * train_frac))
     train = [seqs[i] for i in indices[:n_train]]
     eval_ = [seqs[i] for i in indices[n_train:]]
     return train, eval_
