+++ scripts/run_ple2_semantic_values_3seed.py
@@ -90,7 +90,7 @@ def split_sequences(seqs, train_frac, seed):
     rng = random.Random(seed)
     indices = list(range(len(seqs)))
     rng.shuffle(indices)
-    n_train = max(1, int(round(len(seqs) * train_frac)))
+    n_train = max(1, round(len(seqs) * train_frac))
     return [seqs[i] for i in indices[:n_train]], [seqs[i] for i in indices[n_train:]]
